requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.0.1
openai==1.59.3
pillow==11.0.0
aiofiles==24.1.0
bcrypt>=4.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
slowapi>=0.1.9

# Billing
stripe>=11.0.0

# Testing
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
